+++ apps/dialtone-documentation/thumbs/emoji-text-wrapper.vue
@@ -1,0 +1,7 @@
+<template>
+  <dt-text kind="body" :size="300">
+    <dt-emoji-text-wrapper>
+      Inline emojis like :smile:, :cry:, and 😄.
+    </dt-emoji-text-wrapper>
+  </dt-text>
+</template>