+++ apps/dialtone-documentation/thumbs/mode-island.vue
@@ -1,0 +1,19 @@
+<template>
+  <dt-stack direction="row" gap="200">
+    <dt-box v-dt-mode:invert surface="moderate" padding="200">
+      <dt-text tone="primary">
+        Invert
+      </dt-text>
+    </dt-box>
+    <dt-box v-dt-mode:dark surface="moderate" padding="200">
+      <dt-text tone="primary">
+        Dark
+      </dt-text>
+    </dt-box>
+    <dt-box v-dt-mode:light surface="moderate" padding="200">
+      <dt-text tone="primary">
+        Light
+      </dt-text>
+    </dt-box>
+  </dt-stack>
+</template>