+++ packages/eslint-plugin-dialtone/tests/lib/rules/recommend-typography-style.js
@@ -17,9 +17,10 @@ const rule = require("../../../lib/rules/recommend-typography-style"),
 //------------------------------------------------------------------------------
 
 const ruleTester = new RuleTester({
-  // eslint-disable-next-line n/no-extraneous-require
-  parser: require.resolve('vue-eslint-parser'),
-  parserOptions: { ecmaVersion: 'latest' }
+  languageOptions: {
+    parser: require('vue-eslint-parser'),
+    parserOptions: { ecmaVersion: 'latest' },
+  },
 });
 
 ruleTester.run("recommend-typography-style", rule, {